+++ apps/cowswap-frontend/src/modules/permit/hooks/useGetCachedPermit.ts
@@ -16,6 +16,8 @@ export function useGetCachedPermit(): (
   customSpender?: string,
 ) => Promise<PermitHookData | undefined> {
   const { chainId, account } = useWalletInfo()
+  // TODO M-6 COW-573
+  // This flow will be reviewed and updated later, to include a wagmi alternative
   const provider = useWalletProvider()
   const getCachedPermit = useSetAtom(getPermitCacheAtom)