+++ apps/cowswap-frontend/src/modules/wallet/pure/StatusIcon/index.tsx
@@ -1,4 +1,5 @@
-import { Identicon, ConnectionType } from '@cowprotocol/wallet'
+import { LAUNCH_DARKLY_VIEM_MIGRATION } from '@cowprotocol/common-const'
+import { Identicon, ConnectionType, ConnectorType } from '@cowprotocol/wallet'
 import { CoinbaseWalletIcon, WalletConnectIcon } from '@cowprotocol/wallet'
 
 import { t } from '@lingui/core/macro'
@@ -17,7 +18,7 @@ const IconWrapper = styled.div<{ size?: number }>`
 `
 
 export interface StatusIconProps {
-  connectionType: ConnectionType
+  connectionType?: ConnectionType | string
   account?: string
   size?: number
 }
@@ -25,6 +26,23 @@ export interface StatusIconProps {
 // TODO: Add proper return type annotation
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function StatusIcon({ connectionType, account, size = 16 }: StatusIconProps) {
+  if (LAUNCH_DARKLY_VIEM_MIGRATION) {
+    let image
+    switch (connectionType) {
+      case ConnectorType.COINBASE_WALLET:
+        image = <img src={CoinbaseWalletIcon} alt={t`Coinbase Wallet`} />
+        break
+      case ConnectorType.INJECTED:
+        image = <Identicon account={account} />
+        break
+      case ConnectorType.WALLET_CONNECT_V2:
+        image = <img src={WalletConnectIcon} alt={t`WalletConnect`} />
+        break
+    }
+
+    return <IconWrapper size={size}>{image}</IconWrapper>
+  }
+
   let image
   switch (connectionType) {
     case ConnectionType.INJECTED: