+++ libs/wallet-provider/src/hooks/useWalletChainId.ts
@@ -1,7 +1,15 @@
+import { LAUNCH_DARKLY_VIEM_MIGRATION } from '@cowprotocol/common-const'
 import { useWeb3React } from '@web3-react/core'
 
-export function useWalletChainId(): number | undefined {
-  const { chainId } = useWeb3React()
+import { useConnection } from 'wagmi'
 
-  return chainId
+export function useWalletChainId(): number | undefined {
+  const { chainId: wagmiChainId } = useConnection()
+  const { chainId: web3ChainId } = useWeb3React()
+
+  if (LAUNCH_DARKLY_VIEM_MIGRATION) {
+    return wagmiChainId
+  }
+
+  return web3ChainId
 }